+++ c-core/lrt.h
@@ -3,7 +3,7 @@
 
 #include "compiler.h"
 
-DCP_CONST double lrt(double null_loglik, double alt_loglik)
+DCP_CONST float lrt(float null_loglik, float alt_loglik)
 {
   return -2 * (null_loglik - alt_loglik);
 }
